% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bxcxt.R
\name{bxcxt}
\alias{bxcxt}
\title{Boxcox transform for a response variable}
\usage{
bxcxt(y, x)
}
\arguments{
\item{y}{A vector of response variable}

\item{x}{A vector of explanatory variable}
}
\description{
Boxcox transform for a response variable
}
\examples{
vehicles_obs <- vehicles[vehicles@data$obs1pred0 == 1,]
vehicles_obs$transheavy <- bxcxt(vehicles_obs$heavy, vehicles_obs$width)[[1]]
qqnorm(vehicles_obs$heavy)
qqnorm(vehicles_obs$transheavy)
lambda <- bxcxt(vehicles_obs$heavy, vehicles_obs$width)[[2]]
}
