% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wheatbelt}
\alias{wheatbelt}
\title{Spatial dataset of study area}
\format{\code{wheatbelt}: A Spatial Polygon Data Frame.}
\usage{
wheatbelt
}
\description{
The "wheatbelt" dataset is spatial polylon of the study area
Wheatbelt region, Western Australia (WA), Australia.
More details and data sources can be referred in the Citation Info of the package.
}
\author{
Yongze Song \email{yongze.song@postgrad.curtin.edu.au}
}
\keyword{dataset}
\keyword{wheatbelt}
